IA-787 refactor utils functions, add ModalSubTitle
- move functions used to format options to datasources/utils
- move convertFormStateToDict to /Iaso/utils/forms
